+++ packages/webview_flutter/webview_flutter/lib/webview_flutter.dart
@@ -6,8 +6,11 @@ library webview_flutter;
 
 export 'package:webview_flutter_platform_interface/webview_flutter_platform_interface.dart'
     show
+        JavaScriptAlertDialogRequest,
+        JavaScriptConfirmDialogRequest,
         JavaScriptMessage,
         JavaScriptMode,
+        JavaScriptTextInputDialogRequest,
         LoadRequestMethod,
         NavigationDecision,
         NavigationRequest,